defect #1060041 Add dismiss in context menu for comments in my work

+++ octane-eclipse-plugin/src/com/hpe/octane/ideplugins/eclipse/ui/mywork/job/DismissItemJob.java
@@ -18,6 +18,7 @@ import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 
 import com.hpe.adm.nga.sdk.model.EntityModel;
+import com.hpe.adm.octane.ideplugins.services.CommentService;
 import com.hpe.adm.octane.ideplugins.services.mywork.MyWorkService;
 import com.hpe.octane.ideplugins.eclipse.Activator;
 
@@ -26,6 +27,8 @@ public class DismissItemJob extends Job {
     MyWorkService myWorkService = Activator.getInstance(MyWorkService.class);
     EntityModel entityModel;
     private boolean wasRemoved = false;
+   
+    private CommentService commentService = Activator.getInstance(CommentService.class);
 
     public DismissItemJob(String name, EntityModel entityModel) {
         super(name);
@@ -35,7 +38,12 @@ public class DismissItemJob extends Job {
     @Override
     protected IStatus run(IProgressMonitor monitor) {
         monitor.beginTask(getName(), IProgressMonitor.UNKNOWN);
-        wasRemoved = myWorkService.removeFromMyWork(entityModel);
+        if (entityModel.getType().equals("comment")) {
+        	wasRemoved = commentService.dismissComment(entityModel);
+        } 
+        else {
+        	wasRemoved = myWorkService.removeFromMyWork(entityModel);
+        }
         monitor.done();
         return Status.OK_STATUS;
     }
